Revert "refactor(mtp): extract BaseMTPModel mixin shared by existing MTP draft models (#1337)"
This reverts commit e03ef9a6f597d0b140e1b33e336fa976aadb19bd.
